+++ backend/src/app/models/query_core.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel
 class Rule(BaseModel):
     """Rule model."""
 
-    type: Literal["must_return", "may_return", "max_length"]
+    type: Literal["must_return", "may_return", "max_length", "resolve_entity"]
     options: Optional[List[str]] = None
     length: Optional[int] = None