+++ frontend/src/components/kt/kt-controls/kt-global-rules.tsx
@@ -125,8 +125,8 @@ export function KTGlobalRules(props: BoxProps) {
                 must_return,"red,blue,green",Color
                 <br />
                 max_length,3,
-                {/* <br />
-                replace,"blue,ultramarine,sapphire",Color */}
+                <br />
+                resolve_entity,"blue,ultramarine,sapphire",Color
               </Code>
             </Box>
           </Group>
@@ -247,8 +247,7 @@ export function KTGlobalRules(props: BoxProps) {
 
 const csvJsonSchema = z.array(
   z.object({
-    // rule_type: z.enum(["must_return", "may_return", "max_length", "replace"]),
-    rule_type: z.enum(["must_return", "may_return", "max_length"]),
+    rule_type: z.enum(["must_return", "may_return", "max_length", "resolve_entity"]),
     value: z.string(),
     entity_type: z.string().optional()
   })