+++ frontend/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Allow query without source data if mention is present
-- Removed replace rule as an option for now
+- Removed "resolve entity" rule as an option for now
 
 ## [v0.1.5] - 2024-10-29
 
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updating backend to work with new UI
 - UI: Global rules
 - UI: Reactgrid integration
-- UI: "Replace" rule
+- UI: "Resolve entity" rule
 - UI: Multi-table support
 
 ## [v0.1.4] - 2024-10-16